Risk, Chopsticks, DLD
Added some more things to the risk programs, like determining the probabilites of winning and a program to generate the closed form of some levels of the recursive functions for comparison with my derived functions.

Started work on some more analysis of the Chopsticks game.  I've moved away from the graphical investigation and into the purely programmatic by trying to determine certain properties about the game.

Added several different DLD files for various things.  The biggest program is to try to determine a minimum solution for the BCD 7-segment display problem.