Completed Testing and extension
Completed my testing of the second player winning strategy and determined that if played correctly, the second player can always win.  Furthermore, I extended the program to handle a different number of fingers and hands.  I performed some analysis on these variants and recorded the variants in which a win can be forced and the maximum number of moves it takes to force a win.  From here I plan to do some mathematical analysis to determine exactly what variants allow for a win to be forced.  Additionally, I plan to try to find a simple, memorizable version of the original second player winning strategy.