Additional Chopstick Testing
Performed some more testing with the chopstick analysis program and updated it to make it faster and more memory efficient.  This testing dealt mostly with the variants involving a different number of hands and/or fingers.  I also added a carry-over game-play option.  In this game variant, a hand is only killed by giving it exactly the finger limit; if it goes over, the number of fingers wraps around again, so with a 5-finger limit, 3 hitting 3 would end up with 1 (6-5), instead of 0.  This game variant primarily changes the forced win nature of the (2,5) game, allowing either player to win from the start.  This game variant can also be combined with the allowed stalling variant, although it is rather unexciting.